// apps/fe/enoch/src/constants/onboarding-enums.ts
export enum OnBoardingScreen {
  ChooseAccountType = "ChooseAccountType",
  FillPersonalInformation = "FillPersonalInformation",
  ChooseIndustryType = "ChooseIndustryType",
  ChooseInteresedThings = "ChooseInteresedThings",
  SelectUsersToFollow = "SelectUsersToFollow",
  SelectInterestedCommunities = "SelectInterestedCommunities",
  SubscribeLatestChannels = "SubscribeLatestChannels",
  ChooseAvatar = "ChooseAvatar",
  CreateProfile = "CreateProfile",
  AcceptTermsOfService = "AcceptTermsOfService",
  OnboardingCompleted = "OnboardingCompleted",
  FeaturePopupScreen = "FeaturePopupScreen",
  WelcomeScreenAfterOnboarding = "WelcomeScreenAfterOnboarding",
  UpdateDetailsAfterOnBoardingCompletion = "UpdateDetailsAfterOnBoardingCompletion"
}
export enum GenderEnum {
  MALE = "MALE",
  FEMALE = "FEMALE",
  OTHERS = "OTHERS"
}

export enum RelationshipStatus {
  SINGLE = "SINGLE",
  MARRIED = "MARRIED",
  WIDOWED = "WIDOWED",
  DIVORCED = "DIVORCED",
  DOMESTIC = "DOMESTIC",
  PREFER_NOT_TO_SAY = "PREFER_NOT_TO_SAY"
}

export enum ReasonHereFor {
  SELLING_DIGITAL_ARTS = "SELLING_DIGITAL_ARTS",
  MAKING_SOCIAL_CIRCLE = "MAKING_SOCIAL_CIRCLE",
  INVESTING_IN_MARKETPLACE = "INVESTING_IN_MARKETPLACE",
  PREFER_NOT_TO_SAY = "PREFER_NOT_TO_SAY"
}

export enum AccountType {
  Individual = "Individual",
  Business = "Business"
}

export enum DBModelNames {
  _2FA = "_2FA",
  USER = "User",
  ALLLOWED_USER = "AllowedUser",
  INDUSTRY = "Industry",
  NEWS_CHANNEL = "NewsChannel",
  COMMUNITY = "Community",
  AVATAR = "Avatar",
  BackgroundImage = "BackgroundImage"
}

export enum AvatarCardType {
  SILVER = "SILVER",
  GOLD = "GOLD",
  DIAMOND = "DIAMOND"
}